docs(minesweeper): split modules, apply EXAMPLE_STANDARD, pin cougr-core 1.1.0
Splits the monolithic lib.rs into components.rs (board/mine-layout/
game-state components) and systems.rs (deterministic mine placement,
completion check) per EXAMPLE_STANDARD.md §3. Rewrites README.md to
cover all eight §4 sections, marks the example transitional, bumps
cougr-core to 1.1.0, and commits Cargo.lock. No game rule logic was
changed.

Refs #223
